bmps/flask/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
flask-cors==4.0.0
pymysql==1.1.0
gunicorn==21.2.0
sqlalchemy==2.0.25 